Check if a point is contained in a polygon

based on: https://en.wikipedia.org/wiki/Even

is_point_in_path(x, y, poly)

is_point_df_in_path(point_df, poly_df)

Arguments

x
The x coordinates of point.

y
The y coordinates of point.

poly
a list of numeric vectors defining the polygon list(c(x, y), c(x, y), ...) or a data.frame data.frame(x = c(0,0,1), y = c(0,1,1))

point_df
a data.frame defining the points to check, column names are expected

poly_df
a data.frame defining a polygon, column are expected

Value

Functions

is_point_df_in_path: Find wether points are inside a polygon, defining both as data frames with matching names

Examples

poly_selection <- data.frame(
  Petal.Length = c(2, 4, 6),
  Sepal.Width = c(2, 4, 2)
)

dplyr::mutate(
  iris,
  point_in_poly = is_point_in_path(
    Petal.Length,
    Sepal.Width,
    poly_selection
  )
)

#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species point_in_poly
#> 1            5.1         3.5          1.4         0.2     setosa         FALSE
#> 2            4.9         3.0          1.4         0.2     setosa         FALSE
#> 3            4.7         3.2          1.3         0.2     setosa         FALSE
#> 4            4.6         3.1          1.5         0.2     setosa         FALSE
#> 5            5.0         3.6          1.4         0.2     setosa         FALSE
#> 6            5.4         3.9          1.7         0.4     setosa         FALSE
#> 7            4.6         3.4          1.4         0.3     setosa         FALSE
#> 8            5.0         3.4          1.5         0.2     setosa         FALSE
#> 9            4.4         2.9          1.4         0.2     setosa         FALSE
#> 10           4.9         3.1          1.5         0.1     setosa         FALSE
#> 11           5.4         3.7          1.5         0.2     setosa         FALSE
#> 12           4.8         3.4          1.6         0.2     setosa         FALSE
#> 13           4.8         3.0          1.4         0.1     setosa         FALSE
#> 14           4.3         3.0          1.1         0.1     setosa         FALSE
#> 15           5.8         4.0          1.2         0.2     setosa         FALSE
#> 16           5.7         4.4          1.5         0.4     setosa         FALSE
#> 17           5.4         3.9          1.3         0.4     setosa         FALSE
#> 18           5.1         3.5          1.4         0.3     setosa         FALSE
#> 19           5.7         3.8          1.7         0.3     setosa         FALSE
#> 20           5.1         3.8          1.5         0.3     setosa         FALSE
#> 21           5.4         3.4          1.7         0.2     setosa         FALSE
#> 22           5.1         3.7          1.5         0.4     setosa         FALSE
#> 23           4.6         3.6          1.0         0.2     setosa         FALSE
#> 24           5.1         3.3          1.7         0.5     setosa         FALSE
#> 25           4.8         3.4          1.9         0.2     setosa         FALSE
#> 26           5.0         3.0          1.6         0.2     setosa         FALSE
#> 27           5.0         3.4          1.6         0.4     setosa         FALSE
#> 28           5.2         3.5          1.5         0.2     setosa         FALSE
#> 29           5.2         3.4          1.4         0.2     setosa         FALSE
#> 30           4.7         3.2          1.6         0.2     setosa         FALSE
#> 31           4.8         3.1          1.6         0.2     setosa         FALSE
#> 32           5.4         3.4          1.5         0.4     setosa         FALSE
#> 33           5.2         4.1          1.5         0.1     setosa         FALSE
#> 34           5.5         4.2          1.4         0.2     setosa         FALSE
#> 35           4.9         3.1          1.5         0.2     setosa         FALSE
#> 36           5.0         3.2          1.2         0.2     setosa         FALSE
#> 37           5.5         3.5          1.3         0.2     setosa         FALSE
#> 38           4.9         3.6          1.4         0.1     setosa         FALSE
#> 39           4.4         3.0          1.3         0.2     setosa         FALSE
#> 40           5.1         3.4          1.5         0.2     setosa         FALSE
#> 41           5.0         3.5          1.3         0.3     setosa         FALSE
#> 42           4.5         2.3          1.3         0.3     setosa         FALSE
#> 43           4.4         3.2          1.3         0.2     setosa         FALSE
#> 44           5.0         3.5          1.6         0.6     setosa         FALSE
#> 45           5.1         3.8          1.9         0.4     setosa         FALSE
#> 46           4.8         3.0          1.4         0.3     setosa         FALSE
#> 47           5.1         3.8          1.6         0.2     setosa         FALSE
#> 48           4.6         3.2          1.4         0.2     setosa         FALSE
#> 49           5.3         3.7          1.5         0.2     setosa         FALSE
#> 50           5.0         3.3          1.4         0.2     setosa         FALSE
#> 51           7.0         3.2          4.7         1.4 versicolor          TRUE
#> 52           6.4         3.2          4.5         1.5 versicolor          TRUE
#> 53           6.9         3.1          4.9         1.5 versicolor         FALSE
#> 54           5.5         2.3          4.0         1.3 versicolor          TRUE
#> 55           6.5         2.8          4.6         1.5 versicolor          TRUE
#> 56           5.7         2.8          4.5         1.3 versicolor          TRUE
#> 57           6.3         3.3          4.7         1.6 versicolor         FALSE
#> 58           4.9         2.4          3.3         1.0 versicolor          TRUE
#> 59           6.6         2.9          4.6         1.3 versicolor          TRUE
#> 60           5.2         2.7          3.9         1.4 versicolor          TRUE
#> 61           5.0         2.0          3.5         1.0 versicolor          TRUE
#> 62           5.9         3.0          4.2         1.5 versicolor          TRUE
#> 63           6.0         2.2          4.0         1.0 versicolor          TRUE
#> 64           6.1         2.9          4.7         1.4 versicolor          TRUE
#> 65           5.6         2.9          3.6         1.3 versicolor          TRUE
#> 66           6.7         3.1          4.4         1.4 versicolor          TRUE
#> 67           5.6         3.0          4.5         1.5 versicolor          TRUE
#> 68           5.8         2.7          4.1         1.0 versicolor          TRUE
#> 69           6.2         2.2          4.5         1.5 versicolor          TRUE
#> 70           5.6         2.5          3.9         1.1 versicolor          TRUE
#> 71           5.9         3.2          4.8         1.8 versicolor         FALSE
#> 72           6.1         2.8          4.0         1.3 versicolor          TRUE
#> 73           6.3         2.5          4.9         1.5 versicolor          TRUE
#> 74           6.1         2.8          4.7         1.2 versicolor          TRUE
#> 75           6.4         2.9          4.3         1.3 versicolor          TRUE
#> 76           6.6         3.0          4.4         1.4 versicolor          TRUE
#> 77           6.8         2.8          4.8         1.4 versicolor          TRUE
#> 78           6.7         3.0          5.0         1.7 versicolor         FALSE
#> 79           6.0         2.9          4.5         1.5 versicolor          TRUE
#> 80           5.7         2.6          3.5         1.0 versicolor          TRUE
#> 81           5.5         2.4          3.8         1.1 versicolor          TRUE
#> 82           5.5         2.4          3.7         1.0 versicolor          TRUE
#> 83           5.8         2.7          3.9         1.2 versicolor          TRUE
#> 84           6.0         2.7          5.1         1.6 versicolor          TRUE
#> 85           5.4         3.0          4.5         1.5 versicolor          TRUE
#> 86           6.0         3.4          4.5         1.6 versicolor          TRUE
#> 87           6.7         3.1          4.7         1.5 versicolor          TRUE
#> 88           6.3         2.3          4.4         1.3 versicolor          TRUE
#> 89           5.6         3.0          4.1         1.3 versicolor          TRUE
#> 90           5.5         2.5          4.0         1.3 versicolor          TRUE
#> 91           5.5         2.6          4.4         1.2 versicolor          TRUE
#> 92           6.1         3.0          4.6         1.4 versicolor          TRUE
#> 93           5.8         2.6          4.0         1.2 versicolor          TRUE
#> 94           5.0         2.3          3.3         1.0 versicolor          TRUE
#> 95           5.6         2.7          4.2         1.3 versicolor          TRUE
#> 96           5.7         3.0          4.2         1.2 versicolor          TRUE
#> 97           5.7         2.9          4.2         1.3 versicolor          TRUE
#> 98           6.2         2.9          4.3         1.3 versicolor          TRUE
#> 99           5.1         2.5          3.0         1.1 versicolor          TRUE
#> 100          5.7         2.8          4.1         1.3 versicolor          TRUE
#> 101          6.3         3.3          6.0         2.5  virginica         FALSE
#> 102          5.8         2.7          5.1         1.9  virginica          TRUE
#> 103          7.1         3.0          5.9         2.1  virginica         FALSE
#> 104          6.3         2.9          5.6         1.8  virginica         FALSE
#> 105          6.5         3.0          5.8         2.2  virginica         FALSE
#> 106          7.6         3.0          6.6         2.1  virginica         FALSE
#> 107          4.9         2.5          4.5         1.7  virginica          TRUE
#> 108          7.3         2.9          6.3         1.8  virginica         FALSE
#> 109          6.7         2.5          5.8         1.8  virginica         FALSE
#> 110          7.2         3.6          6.1         2.5  virginica         FALSE
#> 111          6.5         3.2          5.1         2.0  virginica         FALSE
#> 112          6.4         2.7          5.3         1.9  virginica         FALSE
#> 113          6.8         3.0          5.5         2.1  virginica         FALSE
#> 114          5.7         2.5          5.0         2.0  virginica          TRUE
#> 115          5.8         2.8          5.1         2.4  virginica          TRUE
#> 116          6.4         3.2          5.3         2.3  virginica         FALSE
#> 117          6.5         3.0          5.5         1.8  virginica         FALSE
#> 118          7.7         3.8          6.7         2.2  virginica         FALSE
#> 119          7.7         2.6          6.9         2.3  virginica         FALSE
#> 120          6.0         2.2          5.0         1.5  virginica          TRUE
#> 121          6.9         3.2          5.7         2.3  virginica         FALSE
#> 122          5.6         2.8          4.9         2.0  virginica          TRUE
#> 123          7.7         2.8          6.7         2.0  virginica         FALSE
#> 124          6.3         2.7          4.9         1.8  virginica          TRUE
#> 125          6.7         3.3          5.7         2.1  virginica         FALSE
#> 126          7.2         3.2          6.0         1.8  virginica         FALSE
#> 127          6.2         2.8          4.8         1.8  virginica          TRUE
#> 128          6.1         3.0          4.9         1.8  virginica          TRUE
#> 129          6.4         2.8          5.6         2.1  virginica         FALSE
#> 130          7.2         3.0          5.8         1.6  virginica         FALSE
#> 131          7.4         2.8          6.1         1.9  virginica         FALSE
#> 132          7.9         3.8          6.4         2.0  virginica         FALSE
#> 133          6.4         2.8          5.6         2.2  virginica         FALSE
#> 134          6.3         2.8          5.1         1.5  virginica          TRUE
#> 135          6.1         2.6          5.6         1.4  virginica         FALSE
#> 136          7.7         3.0          6.1         2.3  virginica         FALSE
#> 137          6.3         3.4          5.6         2.4  virginica         FALSE
#> 138          6.4         3.1          5.5         1.8  virginica         FALSE
#> 139          6.0         3.0          4.8         1.8  virginica          TRUE
#> 140          6.9         3.1          5.4         2.1  virginica         FALSE
#> 141          6.7         3.1          5.6         2.4  virginica         FALSE
#> 142          6.9         3.1          5.1         2.3  virginica         FALSE
#> 143          5.8         2.7          5.1         1.9  virginica          TRUE
#> 144          6.8         3.2          5.9         2.3  virginica         FALSE
#> 145          6.7         3.3          5.7         2.5  virginica         FALSE
#> 146          6.7         3.0          5.2         2.3  virginica         FALSE
#> 147          6.3         2.5          5.0         1.9  virginica          TRUE
#> 148          6.5         3.0          5.2         2.0  virginica         FALSE
#> 149          6.2         3.4          5.4         2.3  virginica         FALSE
#> 150          5.9         3.0          5.1         1.8  virginica         FALSE